requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
google-genai==1.20.0